requirements.txt
numpy
faiss-cpu
PyPDF2
nltk
sentence-transformers
transformers
torch
